+++ source/ProcessManager.Orchestrations/Program.cs
@@ -86,7 +86,8 @@ OrchestrationDescription CreateDescription_Brs_021_ElectricalHeatingCalculation_
         canBeScheduled: true,
         functionName: nameof(Orchestration_Brs_021_ElectricalHeatingCalculation_V1));
 
-    description.RecurringCronExpression = "0 12,17 * * *"; // Runs at 12:00 and 17:00 every day
+    // Runs at 12:00 and 17:00 every day
+    description.RecurringCronExpression = "0 12,17 * * *";
 
     foreach (var step in Orchestration_Brs_021_ElectricalHeatingCalculation_V1.Steps)
     {
